Treat Closed HVC wins as non-hvc
This change is quite large and changes the way we treat wins that are
confirmed in a financial year which is not the same as the one the win
was created in.

If the confirmation year doesn't have a HVC for the same campaign id
then that HVC is considered closed.

Wins for closed hvcs should be treated the same as non-hvc wins