Update base MI wins view
This changes base MI wins view to use the helper methods
of date_range_start and date_range_end to filter the queryset.

This is required so we can build on it and allow the user to specify
the dates for filtering using query params.

Now we have one places where this comes from we can parse the query
param date and return that instead if it is set, and default to start
and end of the financial if it is not set.

This commit also fixes a bunch of naive tests that relied on the
incorrect behaviour of not filtering the queryset by today as the end
date if the financial year was the current financial year