System checks for constants
we keep editing constants to add new things but sometimes we make a
mistake and use an ID that is already in use.

I have added a system check that tells us if we do this. it will stop
Django from starting and it'll make the tests fail if the check fails