Fix top non hvc count bug
This fixes a bug with the top_non_hvc endpoints where previously they
were reporting a count much bigger than the total number of entries in
the result list.

This was because the result set was an aggregation of of individual
wins. The count was counting the individual wins that made up the
aggregation.