Pin redis to 2.10.6
The libs redis-py-cluster and django-redis caused redis 3 to be
installed because of their relaxed requirements.

This pins redis to version 2.10.6.

Note: at the moment redis-py-cluster==1.3.6 fixed the issue but
django-redis is still causing redis 3 to be installed unless it
comes after redis-py-cluster in the requirements file which is
not ideal.

Pinning redis explicitly seems a better solution for now.